pyproject.toml
[project]
name = "stargazers"
version = "0.1.0"
description = "GitHub Stargazer Analyzer CLI"
authors = [{ name = "Will McGinnis" }]
dependencies = [
    "httpx",
    "pandas",
    "rich",
    "python-dotenv"
]
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-httpx",
    "ruff",
    "python-dotenv"
]

[project.scripts]
stargazers = "stargazers.cli:main"
forkers = "stargazers.cli:main_forkers"

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20", "SIM", "ARG", "DTZ", "PIE", "Q", "RET", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto" 